+++ Controller/WebPayment/CancelFromWidget.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Monext\Payline\Controller\WebPayment;
+
+class CancelFromWidget extends ReturnFromWidget
+{
+
+}