+++ Helper/Data.php
@@ -2,12 +2,12 @@
 
 namespace Monext\Payline\Helper;
 
-use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Framework\App\Helper\Context;
 use Magento\Framework\Math\Random as MathRandom;
 use Magento\Framework\Serialize\Serializer\Json as Serialize;
 use Monext\Payline\Helper\Constants as HelperConstants;
+use Monext\Payline\PaylineApi\Response\GetWebPaymentDetails as ResponseGetWebPaymentDetails;
 
 class Data extends AbstractHelper
 {
@@ -23,6 +23,7 @@ class Data extends AbstractHelper
      * @var Serialize
      */
     protected $serialize;
+
 
     /**
      * @param Context $context
@@ -115,6 +116,11 @@ class Data extends AbstractHelper
         return $status;
     }
 
+    public function isPaymentQuoteFromPayline(\Magento\Quote\Model\Quote\Payment $payment)
+    {
+        return in_array($payment->getMethod(),HelperConstants::AVAILABLE_WEB_PAYMENT_PAYLINE);
+    }
+
     public function isPaymentFromPayline(\Magento\Sales\Model\Order\Payment $payment)
     {
         return in_array($payment->getMethod(),HelperConstants::AVAILABLE_WEB_PAYMENT_PAYLINE);
@@ -175,5 +181,38 @@ class Data extends AbstractHelper
         return $amount;
     }
 
+
+    /**
+     * @return string
+     */
+    public function getMerchantName()
+    {
+        $merchantName = $this->scopeConfig->getValue(HelperConstants::CONFIG_PATH_PAYLINE_GENERAL_MERCHANT_NAME,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+
+        if(empty($merchantName)) {
+            $merchantName = $this->scopeConfig->getValue(\Magento\Store\Model\Information::XML_PATH_STORE_INFO_NAME,
+                \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        }
+
+        return  preg_replace('/[^A-Z0-9]/', '', strtoupper($merchantName)) ?? 'UNDEFINEDMERCHANTNAME';
+    }
+
+    /**
+     * @param ResponseGetWebPaymentDetails $response
+     * @return mixed
+     */
+    public function getUserMessageForCode(ResponseGetWebPaymentDetails $response)
+    {
+        $resultCode = $response->getResultCode();
+
+        $configPath = HelperConstants::CONFIG_PATH_PAYLINE_ERROR_TYPE . substr($resultCode, 1,1);
+        $errorMessage = $this->scopeConfig->getValue($configPath, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        if(empty($errorMessage)) {
+            $errorMessage = $this->scopeConfig->getValue(HelperConstants::CONFIG_PATH_PAYLINE_ERROR_DEFAULT, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        }
+
+        return !empty($errorMessage) ? $errorMessage : $response->getLongErrorMessage();
+    }
 
 }