+++ Model/PaymentManagement.php
@@ -10,6 +10,7 @@ use Magento\Framework\Data\Collection;
 use Magento\Catalog\Model\ResourceModel\Product\Collection as ProductCollection;
 use Magento\Checkout\Api\PaymentInformationManagementInterface as CheckoutPaymentInformationManagementInterface;
 use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Quote\Api\BillingAddressManagementInterface as QuoteBillingAddressManagementInterface;
 use Magento\Quote\Api\CartRepositoryInterface;
 use Magento\Quote\Api\CartTotalRepositoryInterface;
@@ -166,6 +167,9 @@ class PaymentManagement implements PaylinePaymentManagementInterface
      */
     protected $transactionManager;
 
+    protected $gwpdResponseByToken = [];
+
+
     public function __construct(
         CartRepositoryInterface $cartRepository,
         CartTotalRepositoryInterface $cartTotalRepository,
@@ -315,11 +319,15 @@ class PaymentManagement implements PaylinePaymentManagementInterface
 
     protected function callPaylineApiGetWebPaymentDetails($token)
     {
-        $request = $this->requestGetWebPaymentDetailsFactory->create();
-        $request
-            ->setToken($token);
 
-        return $this->paylineApiClient->callGetWebPaymentDetails($request);
+        if(!isset($this->gwpdResponseByToken[$token])) {
+            $request = $this->requestGetWebPaymentDetailsFactory->create();
+            $request->setToken($token);
+
+            $this->gwpdResponseByToken[$token] = $this->paylineApiClient->callGetWebPaymentDetails($request);
+        }
+
+        return $this->gwpdResponseByToken[$token];
     }
 
     protected function callPaylineApiDoCapture(
@@ -371,14 +379,26 @@ class PaymentManagement implements PaylinePaymentManagementInterface
 
     public function synchronizePaymentWithPaymentGatewayFacade($token, $restoreCartOnError = false)
     {
-        $order = $this->paylineOrderManagement->getOrderByToken($token);
 
+        // IN CASE PAYMENT METHOD IS NOT PAYLINE WE EXIT BEFORE DOING ANYTHING
+        $quote = $this->paylineCartManagement->getCartByToken($token);
+        if($quote && $quote->getId()) {
+            $this->paylineOrderManagement->checkQuotePaymentFromPayline($quote);
+            $response = $this->callPaylineApiGetWebPaymentDetails($token);
+            $transactionData = $response->getTransactionData();
+            // IN CASE THERE IS NO TRANSACTION NO NEED TO CREATE AN ORDER
+            if(empty($transactionData) || empty($transactionData['id'])) {
+                throw new \Exception('Payment is not finalized.');
+            }
+        } else {
+            throw new \Exception('Cannot retrieve valid customer cart.');
+        }
+
+        $order = $this->paylineOrderManagement->getOrderByToken($token);
         if (!$order->getId()) {
             $this->paylineCartManagement->placeOrderByToken($token);
             $order = $this->paylineOrderManagement->getOrderByToken($token);
         }
-        // IN CASE PAYMENT METHOD IS NOT PAYLINE WE EXIT
-        $this->paylineOrderManagement->checkOrderPaymentFromPayline($order);
 
         $logData = [
             'token' => $token,
@@ -396,7 +416,7 @@ class PaymentManagement implements PaylinePaymentManagementInterface
                 $this->paylineCartManagement->restoreCartFromOrder($order);
             }
 
-            throw new \Exception(__($order->getPayment()->getData('payline_response')->getLongErrorMessage() ?? $order->getPayment()->getData('payline_error_message') ?: 'Payment is in error.'));
+            throw new LocalizedException(__($order->getPayment()->getData('payline_user_message') ?? $order->getPayment()->getData('payline_response')->getLongErrorMessage() ?: 'Payment is in error.'));
         }
 
         return $this;
@@ -405,30 +425,34 @@ class PaymentManagement implements PaylinePaymentManagementInterface
     protected function synchronizePaymentWithPaymentGateway(OrderPayment $payment, $token)
     {
         $response = $this->callPaylineApiGetWebPaymentDetails($token);
+        if($payment->getLastTransId()) {
+            $message = __('Tansaction already exist for this order. Payline API call to getWebPaymentDetails with return "%1 : %2" was ignored', $response->getResultCode(), $response->getLongErrorMessage());
+            $payment->getOrder()->addStatusHistoryComment($message);
+            $payment->getOrder()->save();
+            throw new \Exception('Transaction already exists for this order');
+        }
+
         $payment->setData('payline_response', $response);
         $paymentTypeManagement = $this->paymentTypeManagementFactory->create($payment);
         if ($response->isSuccess()) {
             if ($paymentTypeManagement->validate($response, $payment)) {
                 $this->handlePaymentSuccessFacade($response, $payment);
             }
+        } elseif ( $response->isWaitingAcceptance() ) {
+            if ($paymentTypeManagement->validate($response, $payment)) {
+                    // TODO: Need to asssociate a transaction
+                    //$this->handlePaymentWaitingAcceptance($payment, $message);
+                    $this->handlePaymentWaitingAcceptanceFacade($response, $payment);
+            }
         } else {
-            $message = $response->getResultCode() . ' : ' . $response->getShortErrorMessage();
-
-            if ($response->isWaitingAcceptance()) {
-                if ($paymentTypeManagement->validate($response, $payment)) {
-                    $this->handlePaymentWaitingAcceptance($payment, $message);
-                }
-            } elseif ($response->isCanceled()) {
-                $this->flagPaymentAsInError($payment, $message);
+            $this->flagPaymentAsInError($payment, $response);
+            if ($response->isCanceled()) {
                 $paymentTypeManagement->handlePaymentCanceled($payment);
             } elseif ($response->isAbandoned()) {
-                $this->flagPaymentAsInError($payment, $message);
                 $this->handlePaymentAbandoned($payment);
             } elseif ($response->isFraud()) {
-                $this->flagPaymentAsInError($payment, $message);
                 $this->handlePaymentFraud($payment);
             } else {
-                $this->flagPaymentAsInError($payment, $message);
                 $this->handlePaymentRefused($payment);
             }
         }
@@ -470,6 +494,23 @@ class PaymentManagement implements PaylinePaymentManagementInterface
             $message ?? $payment->getData('payline_error_message')
         );
     }
+
+    protected function handlePaymentWaitingAcceptanceFacade(
+        ResponseGetWebPaymentDetails $response,
+        OrderPayment $payment
+    )
+    {
+        $message = $response->getResultCode() . ' : ' . $response->getShortErrorMessage();
+        $this->handlePaymentWaitingAcceptance($payment, $message);
+// TODO: Need to asssociate a transaction
+//        $paymentTypeManagement = $this->paymentTypeManagementFactory->create($payment);
+//        $paymentTypeManagement->handlePaymentWaitingAcceptance($response, $payment);
+//        $this->walletManagement->handleWalletReturnFromPaymentGateway($response, $payment);
+//        $this->paylineOrderManagement->sendNewOrderEmail($payment->getOrder());
+
+        return $this;
+    }
+
 
     protected function handlePaymentAbandoned(OrderPayment $payment, $message = null)
     {
@@ -719,11 +760,25 @@ class PaymentManagement implements PaylinePaymentManagementInterface
         return $this;
     }
 
+    /**
+     * @param OrderPayment $payment
+     * @param ResponseGetWebPaymentDetails|null $response
+     * @return OrderPayment
+     */
     protected function flagPaymentAsInError(OrderPayment $payment, $message = null)
     {
+        $message = null;
+        $response = null;
+        if(!$message) {
+            $response = $payment->getData('payline_response');
+            $message = $response->getResultCode() . ' : ' . $response->getShortErrorMessage();
+        }
+
 
         $payment->setData('payline_in_error', true);
         $payment->setData('payline_error_message', $message);
+        $payment->setData('payline_user_message', $response ? $this->helperData->getUserMessageForCode($response) : $message);
+
         return $payment;
     }
 }